refactor: use ValidationPopover from core
