+++ package.json
@@ -99,7 +99,7 @@
     "@dnd-kit/core": "5.0.3",
     "@dnd-kit/sortable": "6.0.1",
     "@kubernetes/client-node": "0.18.0",
-    "@monokle/components": "0.7.5",
+    "@monokle/components": "0.8.0",
     "@monokle/validation": "0.10.3",
     "@open-policy-agent/opa-wasm": "1.8.0",
     "@reduxjs/toolkit": "1.9.1",
