--- src/components/molecules/ValidationPopover/ValidationPopoverContent.styled.tsx
@@ -1,24 +1,0 @@
-import {Space} from 'antd';
-
-import styled from 'styled-components';
-
-import {Colors} from '@shared/styles/colors';
-
-export const Container = styled(Space)`
-  max-height: 160px;
-  overflow-y: auto;
-  padding-right: 8px;
-`;
-
-export const Line = styled.span`
-  color: ${Colors.whitePure};
-`;
-
-export const Message = styled.span`
-  color: ${Colors.grey7};
-
-  &:hover {
-    text-decoration: underline;
-    cursor: pointer;
-  }
-`;
