--- src/components/molecules/ValidationPopover/index.ts
@@ -1,1 +1,0 @@
-export {default} from './ValidationPopover';
