+++ src/components/molecules/index.ts
@@ -18,5 +18,4 @@ export {default as ResourceRefsIconPopover} from './ResourceRefsIconPopover';
 export {default as SaveResourcesToFileFolderModal} from './SaveResourcesToFileFolderModal';
 export {default as SectionRenderer} from './SectionRenderer';
 export {default as TemplateFormRenderer} from './TemplateFormRenderer';
-export {default as ValidationPopover} from './ValidationPopover';
 export {default as Walkthrough} from './Walkthrough';
